requirements.txt
blinker==1.8.2
branca==0.7.2
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
Flask==3.0.3
folium==0.17.0
hmmlearn==0.3.2
idna==3.8
itsdangerous==2.2.0
Jinja2==3.1.4
joblib==1.4.2
MarkupSafe==2.1.5
numpy==2.1.0
osrm-py==0.5
pandas==2.2.2
python-dateutil==2.9.0.post0
pytz==2024.1
requests==2.32.3
scikit-learn==1.5.1
scipy==1.14.1
six==1.16.0
threadpoolctl==3.5.0
tzdata==2024.1
urllib3==2.2.2
Werkzeug==3.0.4
xyzservices==2024.6.0
